Add recursive isNil check for Optionals.
Turns out that it wasn't a change to how optionals work in Swift 5.4 that caused the encoder not to detect nil but that a return type somewhere changed from `Optional<T>` to `Optional<Optional<T>>` exposing a weakness in the previous check that didn't consider nested optionals.

This fixes that weakness by recursively unwrapping optionals until it encounters a "true nil". You'd think the langue would do this for you, and it does when you know the type, but for some reason `someValue as Any == nil` is always false in Swift.